pyproject.toml
[tool.poetry]
name = "neuralnetworks"
version = "0.1.0"
description = ""
authors = ["Christian Coenen <coenen.christian@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.8.3"
tensorflow = "~=2.3.1"
Pillow = "~=7.2.0"
moviepy = "~=1.0.3"
scipy = "~=1.5"
langdetect = "^1.0.8"
SpeechRecognition = "^3.8.1"
google-cloud-speech = "^2.0.1"

[tool.poetry.dev-dependencies]
black = "~=20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''